+++ scriptcore/src/Tomato/Camera.cs
@@ -1,0 +1,12 @@
+namespace TomatoScript.Tomato {
+
+
+    public class Camera : Object {
+
+        public static Camera main {
+            get =>
+        }
+
+    }
+
+}